examples/requirements.txt
Pillow>=9.0.0
requests>=2.26.0 